+++ README.md
@@ -1,14 +1,8 @@
-# BSK Controller
+# BSK ROS2 MPC Controller (BSK-ROS2-MPC)
 
 This package integrates a Model Predictive Controller (MPC) with [Basilisk astrodynamics simulator](https://hanspeterschaub.info/basilisk/) using [BSK-ROS2-Bridge](https://github.com/Thomas-Chan-2019/srl-ros2-BSK-bridge.git).
 
 The MPC is implemented using the [acados framework](https://github.com/acados/acados).
-
-## Citation
-
-If you use this package in academic work, please cite the following paper:
-
-TODO
 
 ## Setup
 
@@ -16,12 +10,12 @@ This package depends on `acados`. Follow the official [installation guide](https
 
 Clone this repo and the following dependencies into your ROS 2 workspace:
 
-* [bsk\_msgs](https://github.com/E-Krantz/bsk_msgs.git)
+* [bsk-msgs](https://github.com/E-Krantz/bsk-msgs.git)
 
 Build the workspace:
 
 ```bash
-colcon build --packages-up-to bsk_controller
+colcon build --packages-up-to bsk-ros2-mpc
 source install/local_setup.bash
 ```
 
@@ -30,7 +24,7 @@ source install/local_setup.bash
 
 ### Launch File Options
 
-`bsk_mpc.launch.py` supports the following arguments:
+`bsk-mpc.launch.py` supports the following arguments:
 
 * 
 
@@ -38,23 +32,23 @@ source install/local_setup.bash
 Basic:
 
 ```bash
-ros2 launch bsk_controller bsk_mpc.launch.py
+ros2 launch bsk-ros2-mpc bsk_mpc.launch.py
 ```
 
 Direct allocation MPC:
 
 ```bash
-ros2 launch bsk_controller bsk_mpc.launch.py type:=da
+ros2 launch bsk-ros2-mpc bsk_mpc.launch.py type:=da
 ```
 
 Wrench MPC:
 
 ```bash
-ros2 launch bsk_controller bsk_mpc.launch.py type:=wrench namespace:=bskSat0
+ros2 launch bsk-ros2-mpc bsk_mpc.launch.py type:=wrench namespace:=bskSat0
 ```
 
 Follower Wrench MPC:
 
 ```bash
-ros2 launch bsk_controller bsk_mpc.launch.py type:=follower_wrench namespace:=bskSat1 name_leader:=bskSat0
+ros2 launch bsk-ros2-mpc bsk_mpc.launch.py type:=follower_wrench namespace:=bskSat1 name_leader:=bskSat0
 ```
