--- bsk_controller/bsk_controller/launch/px4_mpc_leader.launch.py
@@ -1,108 +1,0 @@
-#!/usr/bin/env python
-############################################################################
-#
-#   Copyright (C) 2024 PX4 Development Team. All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions
-# are met:
-#
-# 1. Redistributions of source code must retain the above copyright
-#    notice, this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright
-#    notice, this list of conditions and the following disclaimer in
-#    the documentation and/or other materials provided with the
-#    distribution.
-# 3. Neither the name PX4 nor the names of its contributors may be
-#    used to endorse or promote products derived from this software
-#    without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-# OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
-# AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-#
-############################################################################
-
-__author__ = "Elias Krantz"
-__contact__ = "eliaskra@kth.se"
-
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.substitutions import LaunchConfiguration
-from launch_ros.actions import Node
-
-
-def generate_launch_description():
-    namespace_arg = DeclareLaunchArgument(
-        'namespace',
-        default_value='snap',
-        description='Namespace for all nodes'
-    )
-    type_arg = DeclareLaunchArgument(
-        'type',
-        default_value='wrench',
-        description='Type of the controller (da, wrench, ...)'
-    )
-    period_arg = DeclareLaunchArgument(
-        'period',
-        default_value='20.0',
-        description='Period (seconds) to stay at each waypoint'
-    )
-
-    namespace = LaunchConfiguration('namespace')
-    type = LaunchConfiguration('type')
-    period = LaunchConfiguration('period')
-
-    return LaunchDescription([
-        namespace_arg,
-        type_arg,
-        period_arg,
-
-        # Launch MPC
-        Node(
-            package='bsk_controller',
-            namespace=namespace,
-            executable='bsk_mpc_px4',
-            name='bsk_mpc',
-            output='screen',
-            emulate_tty=True,
-            parameters=[
-                {'type': type}
-            ]
-        ),
-        Node(
-            package='bsk_controller',
-            namespace=namespace,
-            executable='waypoint_publisher',
-            name='waypoint_publisher',
-            output='screen',
-            emulate_tty=True,
-            parameters=[
-                {'period': period},
-                {'is_simulation': False}
-            ],
-        ),
-        Node(
-            package='tf2_ros',
-            executable='static_transform_publisher',
-            name='static_tf_world_to_inertial',
-            arguments=['0', '0', '0', '0', '0', '0', 'map', 'inertial']
-        ),
-        Node(
-                package='tf2_ros',
-                executable='static_transform_publisher',
-                name='static_tf_world_to_camera',
-                #[p, q]
-                # arguments=['0', '0', '2.5', '0', '0.4349655', '0', '0.9', 'world', 'camera_link'] # camera 1
-                arguments=['2', '1.9', '2.3', '0.3010647', '0.3013046', '-0.6395013', '0.6400107', 'map', 'camera_link'] # camera 2
-        ),
-    ])
