--- bsk_controller/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/bsk_controller
-[install]
-install_scripts=$base/lib/bsk_controller
