+++ bsk_ros2_mpc/bsk_planner/simpleWaypoint.py
@@ -1,13 +1,17 @@
+#!/usr/bin/env python
+'''
+Send a single waypoint to a specified namespaced controller
+'''
 import rclpy
 from rclpy.node import Node
 from geometry_msgs.msg import PoseStamped
 import math
 
 # Set these variables directly in the script (angles in degrees)
-namespace = "bskSat1"
-x = 0.0
-y = -3.0
-z = 0.0
+namespace = "bskSat0"
+x = 0.0 # meter
+y = -3.0 # meter
+z = 0.0 # meter
 roll_deg = 0.0   # degrees
 pitch_deg = 0.0 # degrees
 yaw_deg = 0.0    # degrees
