+++ bsk_ros2_mpc/bsk_ros2_mpc/bsk_mpc.py
@@ -1,37 +1,4 @@
 #!/usr/bin/env python
-############################################################################
-#
-#   Copyright (C) 2024 PX4 Development Team. All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions
-# are met:
-#
-# 1. Redistributions of source code must retain the above copyright
-#    notice, this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright
-#    notice, this list of conditions and the following disclaimer in
-#    the documentation and/or other materials provided with the
-#    distribution.
-# 3. Neither the name PX4 nor the names of its contributors may be
-#    used to endorse or promote products derived from this software
-#    without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-# OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
-# AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-#
-############################################################################
-
 __author__ = "Elias Krantz"
 __contact__ = "eliaskra@kth.se"
 
@@ -80,15 +47,15 @@ class BskMpc(Node):
 
         # Create Spacecraft and controller objects
         if self.type == 'da':
-            from bsk_controller.controllers.mpc_da import MpcDa
+            from bsk_ros2_mpc.controllers.mpc_da import MpcDa
             self.mpc = MpcDa()
             self.control = np.zeros((self.mpc.nu, 1))
         elif self.type == 'wrench':
-            from bsk_controller.controllers.mpc_wrench import MpcWrench
+            from bsk_ros2_mpc.controllers.mpc_wrench import MpcWrench
             self.mpc = MpcWrench()
             self.control = np.zeros((self.mpc.nu, 1))
         elif self.type == 'follower_wrench':
-            from bsk_controller.controllers.mpc_follower_wrench import MpcFollowerWrench
+            from bsk_ros2_mpc.controllers.mpc_follower_wrench import MpcFollowerWrench
             self.mpc = MpcFollowerWrench(1) # Assuming 1 other agent (the leader)
             self.control = np.zeros((self.mpc.nu, 1))
