+++ bsk_ros2_mpc/bsk_ros2_mpc/bsk_planner/waypoint_publisher.py
@@ -1,3 +1,5 @@
+
+#!/usr/bin/env python
 import rclpy
 from rclpy.parameter import Parameter
 from rclpy.node import Node
