+++ bsk_ros2_mpc/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>bsk_controller</name>
+  <name>bsk-ros2-mpc</name>
   <version>0.1.0</version>
   <description>Controller node for Basilisk spacecraft</description>
   <maintainer email="eliaskra@kth.se">Elias Krantz</maintainer>
