+++ bsk_ros2_mpc/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/bsk-ros2-mpc
+[install]
+install_scripts=$base/lib/bsk-ros2-mpc
