--- mpc_msgs/srv/SetPose.srv
@@ -1,3 +1,0 @@
-geometry_msgs/Pose pose
----
-bool result
